+++ launch/visualize_franka_duo.launch.py
@@ -23,20 +23,20 @@ from launch.actions import DeclareLaunchArgument, OpaqueFunction
 from launch.substitutions import LaunchConfiguration
 
 
-def robot_state_publisher_spawner(context: LaunchContext, arm_id, arm_prefix, load_gripper, ee_id):
-    arm_id_str = arm_id
+def robot_state_publisher_spawner(context: LaunchContext, robot_type, arm_prefix, load_gripper, ee_id):
+    robot_type_str = robot_type
     arm_prefix_str = arm_prefix
     load_gripper_str = context.perform_substitution(load_gripper)
     ee_id_str = context.perform_substitution(ee_id)
     franka_xacro_filepath = os.path.join(
-        get_package_share_directory("franka_description"),
-        "robots",
-        "fr3_duo",
-        "fr3_duo_" + arm_id_str + ".urdf.xacro",
+        get_package_share_directory('franka_description'),
+        'robots',
+        'fr3_duo',
+        'fr3_duo_' + robot_type_str + '.urdf.xacro',
     )
 
-    if arm_id_str == "fixed_structure":
-        robot_description = xacro.process_file(franka_xacro_filepath).toprettyxml(indent="  ")
+    if robot_type_str == 'fixed_structure':
+        robot_description = xacro.process_file(franka_xacro_filepath).toprettyxml(indent='  ')
     else:
         robot_description = xacro.process_file(
             franka_xacro_filepath,
@@ -49,18 +49,18 @@ def robot_state_publisher_spawner(context: LaunchContext, arm_id, arm_prefix, lo
 
     return [
         Node(
-            package="robot_state_publisher",
-            executable="robot_state_publisher",
-            name="robot_state_publisher_" + arm_id_str + "_" + arm_prefix_str,
-            namespace=arm_id_str + "_" + arm_prefix_str,
-            output="screen",
-            parameters=[{"robot_description": robot_description}],
+            package='robot_state_publisher',
+            executable='robot_state_publisher',
+            name='robot_state_publisher_' + robot_type_str + '_' + arm_prefix_str,
+            namespace=robot_type_str + '_' + arm_prefix_str,
+            output='screen',
+            parameters=[{'robot_description': robot_description}],
         ),
         Node(
-                package="joint_state_publisher_gui",
-                executable="joint_state_publisher_gui",
-                name="joint_state_publisher_gui",
-                namespace=arm_id_str + "_" + arm_prefix_str,
+                package='joint_state_publisher_gui',
+                executable='joint_state_publisher_gui',
+                name='joint_state_publisher_gui',
+                namespace=robot_type_str + '_' + arm_prefix_str,
             ),
     ]
 
@@ -72,8 +72,8 @@ def generate_launch_description():
     ee_id_parameter_name = "ee_id"
     ee_id = LaunchConfiguration(ee_id_parameter_name)
 
-    arm_id_parameter_name = "arm_id"
-    arm_id = LaunchConfiguration(arm_id_parameter_name)
+    robot_type_parameter_name = 'robot_type'
+    robot_type = LaunchConfiguration(robot_type_parameter_name)
 
     rviz_file = os.path.join(
         get_package_share_directory("franka_description"),
@@ -90,12 +90,12 @@ def generate_launch_description():
                 function=robot_state_publisher_spawner, args=[component, "", load_gripper, ee_id]
             ))
         else:
-            print("Spawning arm")
-            arm_id = component.split("_")[0]
-            arm_prefix = component.split("_")[1]
+            print('Spawning arm')
+            robot_type = component.split('_')[0]
+            arm_prefix = component.split('_')[1]
             robot_state_publisher_spawner_opaque_function.append(OpaqueFunction(
                 function=robot_state_publisher_spawner,
-                args=[arm_id, arm_prefix, load_gripper, ee_id]
+                args=[robot_type, arm_prefix, load_gripper, ee_id]
             ))
 
     return LaunchDescription(
@@ -113,9 +113,9 @@ def generate_launch_description():
                 "none, franka_hand, cobot_pump",
             ),
             DeclareLaunchArgument(
-                arm_id_parameter_name,
-                description="ID of the type of arm used. Supporter values: "
-                "fer, fr3, fp3, fr3v2, fr3_duo",
+                robot_type_parameter_name,
+                description='ID of the type of arm used. Supporter values: '
+                'fr3_duo',
             ),
             *robot_state_publisher_spawner_opaque_function,
             Node(